Transmit Mode

+++ jni/tnr/tnr_bch.c
@@ -62,7 +62,7 @@ char api_clisock [DEF_BUF] = DEF_API_CLISOCK;
     // API start/stop
     // Functions called from this chip specific code to generic code:
 
-  int ms_sleep (int ms);
+  long ms_sleep (long ms);
 
   #define  loge(...)  fm_log_print(ANDROID_LOG_ERROR, LOGTAG,__VA_ARGS__)
   #define  logd(...)  fm_log_print(ANDROID_LOG_DEBUG, LOGTAG,__VA_ARGS__)
